Rename path to /4844, move useIntervalTimer to util hooks
