+++ packages/synapse-interface/components/_Transaction/_Transactions.tsx
@@ -3,7 +3,7 @@ import { use_TransactionsState } from '@/slices/_transactions/hooks'
 import { _TransactionDetails } from '@/slices/_transactions/reducer'
 import { _Transaction } from './_Transaction'
 import { checkTransactionsExist } from '@/utils/checkTransactionsExist'
-import { useIntervalTimer } from './helpers/useIntervalTimer'
+import { useIntervalTimer } from '../../utils/hooks/useIntervalTimer'
 
 /** TODO: Update naming once refactoring of previous Activity/Tx flow is done */
 export const _Transactions = ({
