+++ packages/synapse-interface/utils/hooks/useIntervalTimer.ts
